新規でコメントした人のデータが読み込まれない問題を修正
